Adding board config fixtures

+++ app/src/main/java/fr/dmconcept/bob/BobApplication.java
@@ -3,9 +3,13 @@ package fr.dmconcept.bob;
 import android.app.Application;
 import android.database.sqlite.SQLiteDatabase;
 import android.util.Log;
+import fr.dmconcept.bob.models.BoardConfig;
+import fr.dmconcept.bob.models.ServoConfig;
 import fr.dmconcept.bob.models.dao.BoardConfigDao;
 import fr.dmconcept.bob.models.dao.ProjectsDao;
 import fr.dmconcept.bob.models.helpers.BobSqliteOpenHelper;
+
+import java.util.ArrayList;
 
 public class BobApplication extends Application {
 
@@ -20,10 +24,12 @@ public class BobApplication extends Application {
     @Override
     public void onCreate() {
 
-        // DEBUG
         Log.i(TAG, "onCreate()");
-        deleteDatabase("project.db");
-        // DEBUG
+
+        if (BuildConfig.DEBUG) {
+            Log.i(TAG, "DEBUG MODE - Delete databse");
+            deleteDatabase("project.db");
+        }
 
         mOpenHelper = new BobSqliteOpenHelper(this);
         mDatabase   = mOpenHelper.getWritableDatabase();
@@ -31,8 +37,38 @@ public class BobApplication extends Application {
         mBoardConfigDao = new BoardConfigDao(mDatabase);
         mProjectsDao    = new ProjectsDao   (mDatabase, mBoardConfigDao);
 
+        if (BuildConfig.DEBUG)
+            createFixtures();
+
         super.onCreate();
 
+    }
+
+    private void createFixtures() {
+
+        Log.i(TAG, "DEBUG MODE - Creating servos config fixtures");
+
+        ArrayList<ServoConfig> servoConfigs1 = new ArrayList<ServoConfig>();
+        servoConfigs1.add(new ServoConfig(3, 1200, 1660, 50));
+        mBoardConfigDao.save(new BoardConfig("Demo port 3"         , servoConfigs1));
+
+        ArrayList<ServoConfig> servoConfigs2 = new ArrayList<ServoConfig>();
+        servoConfigs2.add(new ServoConfig(4, 1200, 1660, 50));
+        servoConfigs2.add(new ServoConfig(3, 1200, 1660, 50));
+        mBoardConfigDao.save(new BoardConfig("Demo port 3, 4 and 5", servoConfigs2));
+
+        Log.i(TAG, "DEBUG MODE - Creating project...");
+
+        Log.i(TAG, "DEBUG MODE - Fixtures creation done");
+
+    }
+
+    public ProjectsDao getProjectsDao() {
+        return mProjectsDao;
+    }
+
+    public BoardConfigDao getBoardConfigDao() {
+        return mBoardConfigDao;
     }
 
     @Override
@@ -46,12 +82,5 @@ public class BobApplication extends Application {
 
     }
 
-    public ProjectsDao getProjectsDao() {
-        return mProjectsDao;
-    }
-
-    public BoardConfigDao getBoardConfigDao() {
-        return mBoardConfigDao;
-    }
 
 }
